backend/python/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.0
python-dotenv==0.19.0
requests==2.26.0
scikit-learn==1.0.2
numpy==1.21.5
pandas==1.4.2
joblib==1.1.0 